+++ src/utils/client/client.ts
@@ -1,6 +1,6 @@
 import { Client, ClientOptions, Collection } from 'discord.js';
-import BaseEvent from '../utils/structures/BaseEvent';
-import BaseCommand from '../utils/structures/BaseCommand';
+import BaseEvent from '../structures/BaseEvent';
+import BaseCommand from '../structures/BaseCommand';
 
 export default class DiscordClient extends Client {
